# [PERF] account, l10n_*: Improve batch send and print performances