--- a/addons/account/models/account_move.py
+++ b/addons/account/models/account_move.py
@@ -4888,9 +4888,9 @@ class AccountMove(models.Model):
         if self.invoice_pdf_report_id:
             attachments = self.env['account.move.send']._get_invoice_extra_attachments(self)
         else:
-            content, _ = self.env['ir.actions.report']._render('account.account_invoices', self.ids, data={'proforma': True})
+            content, _ = self.env['ir.actions.report']._pre_render_qweb_pdf('account.account_invoices', self.ids, data={'proforma': True})
             attachments = self.env['ir.attachment'].new({
-                'raw': content,
+                'raw': content[self.id],
                 'name': self._get_invoice_proforma_pdf_report_filename(),
                 'mimetype': 'application/pdf',
                 'res_model': self._name,