--- a/addons/account/models/ir_actions_report.py
+++ b/addons/account/models/ir_actions_report.py
@@ -47,7 +47,7 @@ class IrActionsReport(models.Model):
     def _is_invoice_report(self, report_ref):
         return self._get_report(report_ref).report_name in ('account.report_invoice_with_payments', 'account.report_invoice')
 
-    def _render_qweb_pdf(self, report_ref, res_ids=None, data=None):
+    def _pre_render_qweb_pdf(self, report_ref, res_ids=None, data=None):
         # Check for reports only available for invoices.
         # + append context data with the display_name_in_footer parameter
         if self._is_invoice_report(report_ref):
@@ -58,7 +58,19 @@ class IrActionsReport(models.Model):
             if any(x.move_type == 'entry' for x in invoices):
                 raise UserError(_("Only invoices could be printed."))
 
-        return super()._render_qweb_pdf(report_ref, res_ids=res_ids, data=data)
+        content, report_type = super()._pre_render_qweb_pdf(report_ref, res_ids=res_ids, data=data)
+
+        if self._is_invoice_report(report_ref):
+            if report_type == 'html':
+                report = self._get_report(report_ref)
+                bodies, res_ids, *_unused = self._prepare_html(content, report_model=report.model)
+                return {res_id: str(body).encode() for res_id, body in zip(res_ids, bodies)}, 'html'
+            elif report_type == 'pdf':
+                pdf_dict = {res_id: stream['stream'].getvalue() for res_id, stream in content.items()}
+                for stream in content.values():
+                    stream['stream'].close()
+                return pdf_dict, 'pdf'
+        return content, report_type
 
     @api.ondelete(at_uninstall=False)
     def _unlink_except_master_tags(self):