@robodoo 2024-06-03
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/167729).