@aboo-odoo 2024-06-03
> would you mind referring to each PR in each commit message ? 

Done :+1: 

> Would there be a way to ensure that everything is still working correctly ? Or can you explain to me why my worries would be mistaken ?

As we discussed, I tested with long long invoices and no problem there. I only tested with the DIN5008 layout and it all went smoothly :smile: