@casp-odoo 2024-06-03
In which case I'll leave my dearest Jibouille... I mean @jbw-odoo check if I missed anything else ;)